+++ R/get_series_table.R
@@ -10,6 +10,9 @@
 #' @param year_limit optional number of years to paginate request by. Defaults
 #' to 20, the API request cap when using API key. Requests made without an API
 #' key are capped to 10 years.
+#' @param parse_values optional boolean. If set to `true` (default) it will
+#' attempt to parse the contents of `value` and cast numeric strings as numeric
+#' values. If set to `false` it will keep return a `value` column of strings.
 #' @param ... additional parameters to pass to [`bls_request`]
 #'
 #' @return a tibble of observations
@@ -24,7 +27,10 @@
 #' }
 #'
 
-get_series_table <- function(series_id, api_key, start_year=NA, end_year=NA, year_limit=20, ...){
+get_series_table <- function(
+  series_id, api_key,
+  start_year=NA, end_year=NA, year_limit=20, parse_values=TRUE, ...
+  ){
   if(is.na(api_key)) year_limit <- 10
   #auto-magically paginate results
   if(!is.na(start_year) && (end_year - start_year) >= year_limit ){
@@ -35,6 +41,7 @@ get_series_table <- function(series_id, api_key, start_year=NA, end_year=NA, yea
         api_key = api_key,
         start_year = start_year+x,
         end_year = min(end_year, start_year+(x+(year_limit-1))),
+        parse_values = parse_values,
         ...
       )
     )
@@ -42,7 +49,10 @@ get_series_table <- function(series_id, api_key, start_year=NA, end_year=NA, yea
     return(purrr::reduce(tables, dplyr::union_all))
   }
   #query_series(series_id, start_year, end_year, ...)
-  data_as_table(get_series(series_id, start_year, end_year, api_key=api_key, ...)$data)
+  data_as_table(
+    get_series(series_id, start_year, end_year, api_key=api_key, ...)$data,
+    parse_values
+  )
 }
 
 
@@ -57,6 +67,9 @@ get_series_table <- function(series_id, api_key, start_year=NA, end_year=NA, yea
 #' @param year_limit optional number of years to paginate request by. Defaults
 #' to 20, the API request cap when using API key. Requests made without an API
 #' key are capped to 10 years.
+#' @param parse_values optional boolean. If set to `true` (default) it will
+#' attempt to parse the contents of `value` and cast numeric strings as numeric
+#' values. If set to `false` it will keep return a `value` column of strings.
 #' @param ... additional parameters to pass to [`query_n_series`]
 #'
 #' @return a list of tables
@@ -79,7 +92,10 @@ get_series_table <- function(series_id, api_key, start_year=NA, end_year=NA, yea
 #' }
 #'
 
-get_series_tables <- function(series_ids, api_key, start_year=NA, end_year=NA, year_limit=20, ...){
+get_series_tables <- function(
+  series_ids, api_key,
+  start_year=NA, end_year=NA, year_limit=20, parse_values=TRUE, ...
+){
   if(is.na(api_key)) year_limit <- 10
   #auto-magically paginate results
   if(!is.na(start_year) && (end_year - start_year) >= year_limit ){
@@ -90,6 +106,7 @@ get_series_tables <- function(series_ids, api_key, start_year=NA, end_year=NA, y
         api_key = api_key,
         start_year = start_year+x,
         end_year = min(end_year, start_year+(x+(year_limit-1))),
+        parse_values = parse_values,
         ...
       )
     )
@@ -101,7 +118,7 @@ get_series_tables <- function(series_ids, api_key, start_year=NA, end_year=NA, y
 
   lapply(
     get_n_series(series_ids, api_key, start_year, end_year, ...),
-    function(x) { data_as_table(x[['data']]) }
+    function(x) { data_as_table(x[['data']], parse_values) }
   )
 }
 
@@ -115,6 +132,9 @@ get_series_tables <- function(series_ids, api_key, start_year=NA, end_year=NA, y
 #' @param start_year optional numeric 4-digit year
 #' @param end_year optional numeric 4-digit year
 #' @param tidy optional boolean. Return will use [`tidy_periods()`] if true
+#' @param parse_values optional boolean. If set to `true` (default) it will
+#' attempt to parse the values of requested data series and cast numeric strings
+#' as numeric values. If set to `false` it will retain them as strings.
 #' @param ... additional parameters to pass to [`get_series_tables`]
 #'
 #' @return a tibble of multiple merged time series
@@ -132,8 +152,19 @@ get_series_tables <- function(series_ids, api_key, start_year=NA, end_year=NA, y
 #' )
 #' }
 #'
-get_n_series_table <- function(series_ids, api_key, start_year=NA, end_year=NA, tidy=FALSE, ...){
-  tables <- get_series_tables(series_ids, api_key, start_year, end_year, ...)
+get_n_series_table <- function(
+  series_ids, api_key,
+  start_year=NA, end_year=NA, tidy=FALSE, parse_values=TRUE, ...
+){
+  tables <- get_series_tables(
+    series_ids = series_ids,
+    api_key = api_key,
+    start_year = start_year,
+    end_year = end_year,
+    parse_values = parse_values,
+    ...
+  )
+
   if(tidy){
     return(merge_tidy_tables(lapply(tables, tidy_periods)))
   }